Cambiar la relacion de de los ids en lugar de resivir un entero mappear una cadena

+++ Services/ImagenService.cs
@@ -22,7 +22,7 @@ namespace Grpc_AutorImagen.Services
                 var imagen = new AutorImagen
                 {
                     Contenido = request.Contenido.ToByteArray(),
-                    IdAutorLibro = (int)request.IdAutorLibro
+                    IdAutorLibro = request.IdAutorLibro
                 };
 
                 _contexto.AutoresImagenes.Add(imagen);
